+++ frontend/src/components/molecules/ChartCards/StableCoinDistributionChart.vue
@@ -39,25 +39,19 @@ watchEffect(() => {
 		return
 	}
 
-	setTimeout(() => {
-		const totalSupplySum = stablecoins.reduce((sum, coin) => {
-			const totalSupply = coin.totalSupply ?? 0
-			const decimal = coin.decimal ?? 0
-			return sum + totalSupply / 10 ** decimal
-		}, 0)
+	const totalSupplySum = stablecoins.reduce((sum, coin) => {
+		return sum + (coin.normalizedCurrentSupply ?? 0)
+	}, 0)
 
-		delayedSupplyPercentage.value = stablecoins.map(coin => ({
-			symbol: coin.tokenId || '',
-			supplyPercentage:
-				totalSupplySum > 0
-					? (
-							((coin.totalSupply ?? 0) /
-								10 ** (coin.decimal ?? 0) /
-								totalSupplySum) *
-							100
-					  ).toString()
-					: '0',
-		}))
-	}, 1000)
+	delayedSupplyPercentage.value = stablecoins.map(coin => ({
+		symbol: coin.tokenId || '',
+		supplyPercentage:
+			totalSupplySum > 0
+				? (
+						((coin.normalizedCurrentSupply ?? 0) / totalSupplySum) *
+						100
+				  ).toString()
+				: '0',
+	}))
 })
 </script>