+++ backend/src/migrations.rs
@@ -288,15 +288,17 @@ pub enum SchemaVersion {
     AddTokenModulePauseUnpauseStatus,
     #[display("0044: Normalize metrics for PLT cumulative transfer amount")]
     NormalizePltCumulativeTransferAmount,
+    #[display("0045: Alter PLT tokens add current_supply column")]
+    AlterPltTokensAddCurrentSupplyColumn,
 }
 impl SchemaVersion {
     /// The minimum supported database schema version for the API.
     /// Fails at startup if any breaking (destructive) database schema versions
     /// have been introduced since this version.
     pub const API_SUPPORTED_SCHEMA_VERSION: SchemaVersion =
-        SchemaVersion::AddTokenModulePauseUnpauseStatus;
+        SchemaVersion::AlterPltTokensAddCurrentSupplyColumn;
     /// The latest known version of the schema.
-    const LATEST: SchemaVersion = SchemaVersion::NormalizePltCumulativeTransferAmount;
+    const LATEST: SchemaVersion = SchemaVersion::AlterPltTokensAddCurrentSupplyColumn;
 
     /// Parse version number into a database schema version.
     /// None if the version is unknown.
@@ -362,6 +364,7 @@ impl SchemaVersion {
             SchemaVersion::AlterPltAccountsAddNotNullConstraint => false,
             SchemaVersion::AddTokenModulePauseUnpauseStatus => false,
             SchemaVersion::NormalizePltCumulativeTransferAmount => false,
+            SchemaVersion::AlterPltTokensAddCurrentSupplyColumn => false,
         }
     }
 
@@ -416,6 +419,7 @@ impl SchemaVersion {
             SchemaVersion::AlterPltAccountsAddNotNullConstraint => false,
             SchemaVersion::AddTokenModulePauseUnpauseStatus => false,
             SchemaVersion::NormalizePltCumulativeTransferAmount => false,
+            SchemaVersion::AlterPltTokensAddCurrentSupplyColumn => false,
         }
     }
 
@@ -722,7 +726,15 @@ impl SchemaVersion {
                     .await?;
                 SchemaVersion::NormalizePltCumulativeTransferAmount
             }
-            SchemaVersion::NormalizePltCumulativeTransferAmount => unimplemented!(
+            SchemaVersion::NormalizePltCumulativeTransferAmount => {
+                tx.as_mut().execute(sqlx::raw_sql(
+                    include_str!(
+                        "./migrations/m0045-alter-plt-tokens-current-supply.sql"
+                    )))
+                    .await?;
+                SchemaVersion::AlterPltTokensAddCurrentSupplyColumn
+            }
+            SchemaVersion::AlterPltTokensAddCurrentSupplyColumn => unimplemented!(
                 "No migration implemented for database schema version {}",
                 self.as_i64()
             ),