+++ frontend/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased 
 
+### Fixed 
+
+- Fixed plt dashboard now displays total `10` plt sorted by current supply.
+
 ## [1.7.26] - 2025-09-16
 
 ### Fixed