+++ frontend/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased 
 
+### Fixed
+
+- Fixed totals supply discrepancy between actual data and ccdscan data on plt overview page.
+
+- Fixed plt dashboard now displays total `10` plt sorted by current supply.
+
 ## [1.7.26] - 2025-09-16
 
 ### Fixed